Update after PR comments
