+++ src/__test__/Entry.test.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
-import { mount, shallow } from 'enzyme';
-import { MemoryRouter, Link } from 'react-router-dom';
-import Entry, { History, Related, Badge } from '../js/components/Entry';
+import { shallow } from 'enzyme';
+import Entry from '../js/components/Entry';
 
 const mockEntries = {
   java: {
@@ -24,8 +23,29 @@ const mockEntries = {
     quadrant: { index: 2, dirname: 'dev' },
     filename: 'java.html',
   },
+  lisp: {
+    label: 'Lisp',
+    blip: {
+      quadrant: 2,
+      since: 'Jan 1, 1970',
+      ring: 3,
+      ringName: 'Hold',
+      moved: false,
+      link: '/entries/lisp.html',
+      active: false,
+      history: [
+        { date: 'Jan 1, 1970', ringName: 'Adopt' },
+        { date: 'Mar 20, 2018', ringName: 'Hold' },
+      ],
+    },
+    description: 'Description',
+    rationale: 'Rationale',
+    quadrant: { index: 2, dirname: 'dev' },
+    filename: 'lisp.html',
+  },
   php: {
     name: 'PHP',
+    logo: 'https://www.php.net/images/logos/php-logo.svg',
     blip: {
       label: 'PHP',
       quadrant: 2,
@@ -64,57 +84,17 @@ describe('<Entry />', () => {
     expect(component).toMatchSnapshot();
   });
 
-  describe('<History />', () => {
-    test('It renders history entries', () => {
-      const component = shallow(<History history={mockEntries.php.blip.history} />);
-      expect(component).toMatchSnapshot();
-      expect(component.find('li')).toHaveLength(2);
-    });
-
-    test('It renders nothing for missing history', () => {
-      const component = shallow(<History history={[]} />);
-      expect(component).toMatchSnapshot();
-    });
+  test('It renders logotype if present', () => {
+    const match = { params: { id: 'php' } };
+    const component = shallow(<Entry match={match} />);
+    expect(component.find('img')).toHaveLength(1);
+    expect(component).toMatchSnapshot();
   });
 
-  describe('<Related />', () => {
-    test('It renders related entries', () => {
-      const component = shallow(<Related related={mockEntries.java.related} />);
-      expect(component).toMatchSnapshot();
-      expect(component.find('li')).toHaveLength(1);
-    });
-
-    test('It renders nothing for no relations', () => {
-      const component = shallow(<Related related={[]} />);
-      expect(component).toMatchSnapshot();
-    });
-
-    test('It is possible to click related link', async () => {
-      const component = mount(
-        <MemoryRouter>
-          <Related related={mockEntries.java.related} />
-        </MemoryRouter>,
-      );
-
-      const { history } = component.find(MemoryRouter).instance();
-      expect(history.length).toEqual(1);
-
-      component.find(Link).first().simulate('click', { button: 0 });
-
-      expect(history.length).toEqual(2);
-      expect(history.location.pathname).toEqual('/php.html');
-    });
-  });
-
-  describe('<Badge />', () => {
-    test('It renders badge with icon', () => {
-      const component = shallow(<Badge icon="test" text="test" />);
-      expect(component).toMatchSnapshot();
-    });
-
-    test('It renders badge with extra className', () => {
-      const component = shallow(<Badge icon="test" text="test" className="my-test" />);
-      expect(component).toMatchSnapshot();
-    });
+  test('It renders inactive badge', () => {
+    const match = { params: { id: 'lisp' } };
+    const component = shallow(<Entry match={match} />);
+    expect(component.find('.inactive')).toHaveLength(1);
+    expect(component).toMatchSnapshot();
   });
 });
